refactor: centralize product data and use numeric price
