+++ src/App.test.js
@@ -1,8 +1,13 @@
 import { render, screen } from '@testing-library/react';
 import App from './App';
+import { CartProvider } from './context/CartContext';
 
 test('renders learn react link', () => {
-  render(<App />);
+  render(
+    <CartProvider>
+      <App />
+    </CartProvider>
+  );
   const linkElement = screen.getByText(/learn react/i);
   expect(linkElement).toBeInTheDocument();
 });
